implement confirm msg on delete

+++ src/MyEdo/MyEdo.Web/ClientApp/src/app/components/skill/skill/skill.component.ts
@@ -2,6 +2,10 @@ import { Component, EventEmitter, Input, OnInit, Output } from "@angular/core";
 import SkillCategoryModel from "../../../core/models/skill-model";
 import { SkillService } from "../../../core/services/skill.service";
 import { Router } from "@angular/router";
+import {
+  ConfirmBoxInitializer,
+  DialogLayoutDisplay,
+} from "@costlydeveloper/ngx-awesome-popup";
 
 @Component({
   selector: "app-skill",
@@ -10,17 +14,16 @@ import { Router } from "@angular/router";
 })
 export class SkillComponent implements OnInit {
   clickButton: boolean = false;
+  deletionMsg = "Are you sure that you want to delete the skill category?";
+
   @Input("skill")
   skill: SkillCategoryModel;
   skillCategories: SkillCategoryModel[];
-  
+
   @Output()
   deleteEvent = new EventEmitter();
 
-  constructor(
-    private skillService: SkillService,
-    private router: Router
-  ) {}
+  constructor(private skillService: SkillService, private router: Router) {}
 
   ngOnInit() {
     this.skillService
@@ -35,14 +38,24 @@ export class SkillComponent implements OnInit {
 
   public deleteCategory(categoryId: string) {
     this.clickButton = true;
-    const currentCategory = this.GetCurrentCategory(categoryId);
-    const body = {
-      id: currentCategory.categoryId,
-      name: currentCategory.categoryName,
-    };
 
-    this.skillService.deleteCategory(body).subscribe(() => {
-      this.deleteEvent.emit(null);
+    const confirmBox = new ConfirmBoxInitializer();
+    confirmBox.setTitle(this.deletionMsg);
+    confirmBox.setButtonLabels("YES", "NO");
+
+    const subscription = confirmBox.openConfirmBox$().subscribe((resp) => {
+      if (resp.Success) {
+        const currentCategory = this.GetCurrentCategory(categoryId);
+        const body = {
+          id: currentCategory.categoryId,
+          name: currentCategory.categoryName,
+        };
+
+        this.skillService.deleteCategory(body).subscribe(() => {
+          this.deleteEvent.emit(null);
+        });
+      }
+      subscription.unsubscribe();
     });
   }
